Reorganize the steps, tasks, and results of the  model protocol
